Rename 'fixfloat' to 'floatFormatter'
